+++ src/presentacion/vistas/VistaWidgetRAM.java
@@ -31,7 +31,7 @@ public class VistaWidgetRAM extends JPanel {
     private JButton btnResaltarMAR;
     private JPanel parentPanel;
     private JScrollPane panelMemoria;
-    private byte valorMAR;
+    private int valorMAR;
     private boolean debeResaltarMAR;
 
     private ControladorWidgetRAM control;
@@ -229,11 +229,11 @@ public class VistaWidgetRAM extends JPanel {
         this.btnResaltarMAR = highlightMarButton;
     }
 
-    public byte getValorMAR() {
+    public int getValorMAR() {
         return valorMAR;
     }
 
-    public void setValorMAR(byte v) {
+    public void setValorMAR(int v) {
         this.valorMAR = v;
     }
